+++ daGYM/src/main/java/com/kh/dagym/community/controller/EventController.java
@@ -8,26 +8,27 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import com.kh.dagym.common.PageInfo;
 import com.kh.dagym.community.service.EventService;
+import com.kh.dagym.community.service.EventServiceImpl;
 
 @Controller
 @RequestMapping("/event/*")
 public class EventController {
 	
+	@Autowired
 	private EventService eventService;
 	
 	private final int BOARD_TYPE = 1;
 	
-//	@Autowired
-//	public EventController(EventService eventService) {
-//		super();
-//		this.eventService = eventService;
-//	}
 
 	@GetMapping("list")
 	public String eventList(@RequestParam(value = "cp", required = false, defaultValue = "1") int cp) {
+		long start = System.currentTimeMillis();
 		
-//		PageInfo pInfo = eventService.pagenation(BOARD_TYPE, cp);
+		PageInfo pInfo = eventService.pagenation(BOARD_TYPE,cp);
+		long end = System.currentTimeMillis();
 		
-		return null;
+		System.out.println((end - start)/1000 + "초");
+
+		return "community/eventList";
 	}
 }
