+++ daGYM/src/main/java/com/kh/dagym/community/model/dao/EventDAO.java
@@ -4,7 +4,7 @@ import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-//@Repository
+@Repository
 public class EventDAO {
 	
 	private SqlSessionTemplate sqlSession;
@@ -20,8 +20,8 @@ public class EventDAO {
 	 * @param BOARD_TYPE
 	 * @return listCount
 	 */
-	public int getListCount(int BOARD_TYPE) {
-		return sqlSession.selectOne("communityMapper.getListCount", BOARD_TYPE);
+	public int getListCount() {
+		return sqlSession.selectOne("eventMapper.getListCount");
 	}
 
 }
