+++ daGYM/src/main/resources/mapper/event-mapper.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="communityMapper">
-	<select id="getListCount"></select>
+<mapper namespace="eventMapper">
+	<select id="getListCount" resultType="_int">
+		SELECT COUNT(*) FROM V_EVENT
+		WHERE END_DATE >= SYSDATE
+	</select>
   
 </mapper>
