+++ daGYM/src/main/resources/mybatis-config.xml
@@ -22,5 +22,6 @@
 	<!--  mapper.xml 파일을 등록하는 태그 -->
 	<mappers>
 		<mapper resource="/mapper/member-mapper.xml"/>
+		<mapper resource="/mapper/event-mapper.xml"/>
 	</mappers>
 </configuration>
