+++ daGYM/src/main/webapp/WEB-INF/views/common/header.jsp
@@ -30,7 +30,7 @@
                         <li><a href="#">Home</a></li>
                         <li><a href="#">소개</a></li>
                         <li><a href="${contextPath}/trainer/trainerView">트레이너 예약</a></li>
-                        <li class="active"><a href="#">커뮤니티</a></li>
+                        <li><a href="${contextPath }/event/list">커뮤니티</a></li>
                         <li><a href="${contextPath}/service/questionList/3">고객센터</a></li>
                     </ul>
                 </nav>
