+++ daGYM/src/main/webapp/WEB-INF/views/community/eventList.jsp
@@ -1,0 +1,14 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+</head>
+<body>
+   <%@ include file="../common/header.jsp" %>
+			<jsp:include page="sideMenu.jsp"/>
+   <%@ include file="../common/footer.jsp" %>
+</body>
+</html>
