+++ daGYM/src/main/webapp/WEB-INF/views/community/sideMenu.jsp
@@ -1,0 +1,18 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+</head>
+<body>
+	<div class="col-sm-3 mt-5">
+	<h3>Side Menu</h3>
+	<ul class="list-group">
+		<li class="list-group-item list-group-item-action"><a href="${contextPath}/event/list">이벤트</a></li>
+		<li class="list-group-item list-group-item-action"><a href="changePwd">정보 게시판</a></li>
+	</ul>
+</div>
+</body>
+</html>
