+++ daGYM/src/main/webapp/main.jsp
@@ -48,7 +48,7 @@
                         <li><a href="#">Home</a></li>
                         <li><a href="#">소개</a></li>
                         <li><a href="trainer/trainerView">트레이너 예약</a></li>
-                        <li><a href="#">커뮤니티</a></li>
+                        <li><a href="${contextPath }/event/list">커뮤니티</a></li>
                         <li><a href="#">고객센터</a></li>
                     </ul>
                 </nav>
